HP_A2_02: rename folder, create routeConstants etc.
