+++ src/components/MainLogo/index.ts
