+++ src/components/SearchBar/SearchBar.styles.tsx
