+++ src/constant/routeConstant.ts
@@ -1,0 +1,8 @@
+/**
+ * Various routes used in the application.
+ * @constant
+ */
+export const ROUTES = {
+    ROOT: '/',
+    NOTIFICATIONS: '/notifications',
+};
