+++ src/layouts/RootLayout.tsx
@@ -1,4 +1,4 @@
-import { Header } from 'container';
+import { Header } from 'containers';
 import { Outlet } from 'react-router-dom';
 
 export const RootLayout = () => (
